switched binomial to long

+++ Binomial.java
@@ -13,7 +13,7 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) {
+	public static long binomial1(long n, long k) {
 		if (k > n)
 			return 0;
 		if (k == 0 || n == 0)
@@ -22,8 +22,8 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int[][] memo = new int[n + 1][k + 1];
+	public static long binomial(int n, int k) {
+		long[][] memo = new long[n + 1][k + 1];
 		for (int i = 0; i <= n; i++) {
 			for (int j = 0; j <= k; j++) {
 				memo[i][j] = -1;
@@ -32,7 +32,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
@@ -46,7 +46,7 @@ public class Binomial {
 			memo[n][k] = 1;
 			return 1;
 		}
-		memo[n][k] = binomial(n - 1, k, memo) + binomial(n - 1, k - 1, memo);
+		memo[n][k] = (long)binomial(n - 1, k, memo) + binomial(n - 1, k - 1, memo);
 		return memo[n][k];
 	}
 }
